requirements.txt
azure-cosmosdb-table==1.0.5
flask==1.1.1
